Adds testing for the metric method adds right config with jest.timeout to 10000 in the tests, tests are working and running green
