+++ package.json
@@ -3,10 +3,13 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "enzyme": "^3.8.0",
+    "enzyme-adapter-react-16": "^1.7.1",
     "react": "^16.7.0",
     "react-dom": "^16.7.0",
-    "react-scripts": "2.1.3",
-    "react-undraw-illustrations": "^2.0.3"
+    "react-scripts": "^2.1.3",
+    "react-test-renderer": "^16.7.0",
+    "sinon": "^7.2.2"
   },
   "scripts": {
     "build:css": "postcss src/styles/index.css -o src/index.css",
@@ -28,10 +31,10 @@
   ],
   "devDependencies": {
     "autoprefixer": "^9.4.4",
+    "postcss-cli": "^6.1.1",
+    "tailwindcss": "^0.7.3",
     "jest-dev-server": "^3.7.0",
     "jest-puppeteer": "^3.7.0",
-    "postcss-cli": "^6.1.1",
-    "puppeteer": "^1.11.0",
-    "tailwindcss": "^0.7.3"
+    "puppeteer": "^1.11.0"
   }
 }
